chore: 주기, 이벤트 저장은 consumer(sub)의 역할, 이에 따른 수정
- eventinfo, cycle등의 엔티티를 database 모듈로 통합
- emulator는 pub 혹은 procedure의 역할이므로 비즈니스 로직을 가지지 않음(mq통신용)
- consumer에서 최종적으로 db에 저장하는 내용을 처리
- [database, common ( emulator -> mq -> consumer and domain )] 구조
- consumer는 domain을 알 수 있지만, domain은 consumer을 알 수 없음
