+++ consumer/build.gradle
@@ -1,7 +1,7 @@
 plugins {
-    id 'java'
     id 'org.springframework.boot' version '3.4.5'
     id 'io.spring.dependency-management'
+    id 'java'
 }
 
 group = 'com'
@@ -12,8 +12,15 @@ repositories {
 }
 
 dependencies {
+    implementation 'org.springframework.boot:spring-boot-starter-web'
+    implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
+    testImplementation 'org.springframework.boot:spring-boot-starter-test'
     testImplementation platform('org.junit:junit-bom:5.10.0')
     testImplementation 'org.junit.jupiter:junit-jupiter'
+
+    implementation project(':domain')
+    implementation project(':database')
+    implementation project(':common')
 }
 
 test {
