+++ consumer/src/main/java/com/pickcar/application/CycleService.java
@@ -1,8 +1,8 @@
 package com.pickcar.application;
 
-import com.pickcar.domain.Cycle;
-import com.pickcar.domain.CycleInfo;
-import com.pickcar.domain.EventInfo;
+import com.pickcar.emulator.domain.Cycle;
+import com.pickcar.emulator.domain.CycleInfo;
+import com.pickcar.emulator.domain.EventInfo;
 import com.pickcar.infrastructure.CycleRepository;
 import com.pickcar.presentation.dto.request.CycleStoreRequest;
 import java.time.LocalDateTime;
