+++ consumer/src/main/java/com/pickcar/infrastructure/EmulatorRepository.java
@@ -1,6 +1,6 @@
 package com.pickcar.infrastructure;
 
-import com.pickcar.domain.Terminal;
+import com.pickcar.emulator.domain.Terminal;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface EmulatorRepository extends JpaRepository<Terminal, Long> {
