+++ consumer/src/main/java/com/pickcar/presentation/dto/request/CycleStoreRequest.java
@@ -1,7 +1,7 @@
 package com.pickcar.presentation.dto.request;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
-import com.pickcar.domain.CycleInfo;
+import com.pickcar.emulator.domain.CycleInfo;
 import java.time.LocalDateTime;
 import java.util.List;
 import lombok.AllArgsConstructor;
