+++ database/src/main/java/com/pickcar/emulator/domain/CycleInfoConverter.java
@@ -1,10 +1,9 @@
-package com.pickcar.infrastructure;
+package com.pickcar.emulator.domain;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.JavaType;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
-import com.pickcar.domain.CycleInfo;
 import jakarta.persistence.AttributeConverter;
 import jakarta.persistence.Converter;
 import java.io.IOException;
