+++ domain/build.gradle
@@ -24,7 +24,6 @@ dependencies {
 
     implementation project(':database')       //DB 모듈 의존성 추가
     implementation project(':common')
-    implementation project(':emulator')
 
     //로그 수집
     implementation 'net.logstash.logback:logstash-logback-encoder:8.1'
