+++ domain/src/main/java/com/pickcar/drivehistory/presentation/DriveHistoryController.java
@@ -2,16 +2,11 @@ package com.pickcar.drivehistory.presentation;
 
 import com.pickcar.drivehistory.application.DriveHistoryService;
 import com.pickcar.drivehistory.presentation.dto.response.ExampleResponse;
-import com.pickcar.presentation.dto.response.SuccessResponse;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 
@@ -28,14 +23,5 @@ public class DriveHistoryController implements DriveHistoryApiDocs {
     @GetMapping("/example")
     public ResponseEntity<ExampleResponse> example() {
         return ResponseEntity.ok(new ExampleResponse("example"));       //Fixme: 실제로는 서비스에서 구성되어야 함
-    }
-
-    @PostMapping("/historyTest/{reservationId}")
-    public ResponseEntity<Void> testWrite(@PathVariable Long reservationId) {
-
-        log.info("mapped history test : {}", reservationId);
-        driveHistoryService.write(reservationId);
-
-        return ResponseEntity.ok().build();
     }
 }
