+++ emulator/build.gradle
@@ -19,6 +19,10 @@ dependencies {
     implementation 'mysql:mysql-connector-java:8.0.33'
     testImplementation platform('org.junit:junit-bom:5.10.0')
     testImplementation 'org.junit.jupiter:junit-jupiter'
+
+    implementation project(':domain')
+    implementation project(':database')
+    implementation project(':consumer')
 }
 
 bootJar {
