--- emulator/src/main/java/com/pickcar/domain/Terminal.java
@@ -1,32 +1,0 @@
-package com.pickcar.domain;
-
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Builder
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor(access = AccessLevel.PROTECTED)
-public class Terminal {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    private Long vehicleId;
-
-    private String mdn;
-
-    private String terminalId;
-
-    private String manufactureId;
-
-    private String packetVersion;
-
-    private String deviceId;
-}
