+++ emulator/src/main/java/com/pickcar/presentation/controller/EmulatorController.java
@@ -6,4 +6,5 @@ import org.springframework.web.bind.annotation.RequestMapping;
 @Controller
 @RequestMapping("/emulator")
 public class EmulatorController {
+    //Pub (or Procedure) Controller => 애뮬레이터(K6  <-> MQ) 간의 소통
 }
